pyproject.toml
[tool.poetry]
name = "photoshare"
version = "0.1.0"
description = ""
authors = ["sebastianLedzianowski <136706736+sebastianLedzianowski@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.27"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
pydantic = "^2.6.2"
cloudinary = "^1.38.0"
redis = "^5.0.1"
jose = "^1.0.0"
passlib = "^1.7.4"
fastapi-mail = "^1.4.1"
fastapi-limiter = "^0.1.6"
alembic = "^1.13.1"
python-jose = "^3.3.0"
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"
bcrypt = "^4.1.2"
boto3 = "^1.34.50"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
